Move constants for buffs from root and expose in buffs namespace
