+++ docs/enums/_buffs_constants_.known_passive_id.md
@@ -1,0 +1,17 @@
+[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["buffs/constants"](../modules/_buffs_constants_.md) › [KNOWN_PASSIVE_ID](_buffs_constants_.known_passive_id.md)
+
+# Enumeration: KNOWN_PASSIVE_ID
+
+## Index
+
+### Enumeration members
+
+* [TriggeredEffect](_buffs_constants_.known_passive_id.md#triggeredeffect)
+
+## Enumeration members
+
+###  TriggeredEffect
+
+• **TriggeredEffect**: = "66"
+
+Defined in buffs/constants.ts:6
