+++ docs/enums/_buffs_constants_.known_proc_id.md
@@ -1,0 +1,17 @@
+[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["buffs/constants"](../modules/_buffs_constants_.md) › [KNOWN_PROC_ID](_buffs_constants_.known_proc_id.md)
+
+# Enumeration: KNOWN_PROC_ID
+
+## Index
+
+### Enumeration members
+
+* [BurstHeal](_buffs_constants_.known_proc_id.md#burstheal)
+
+## Enumeration members
+
+###  BurstHeal
+
+• **BurstHeal**: = "2"
+
+Defined in buffs/constants.ts:2
