--- docs/enums/_constants_.known_passive_id.md
@@ -1,17 +1,0 @@
-[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["constants"](../modules/_constants_.md) › [KNOWN_PASSIVE_ID](_constants_.known_passive_id.md)
-
-# Enumeration: KNOWN_PASSIVE_ID
-
-## Index
-
-### Enumeration members
-
-* [TriggeredEffect](_constants_.known_passive_id.md#triggeredeffect)
-
-## Enumeration members
-
-###  TriggeredEffect
-
-• **TriggeredEffect**: = "66"
-
-*Defined in [constants.ts:6](https://github.com/BluuArc/bfmt-utilities/blob/master/src/constants.ts#L6)*
