--- docs/enums/_constants_.known_proc_id.md
@@ -1,17 +1,0 @@
-[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["constants"](../modules/_constants_.md) › [KNOWN_PROC_ID](_constants_.known_proc_id.md)
-
-# Enumeration: KNOWN_PROC_ID
-
-## Index
-
-### Enumeration members
-
-* [BurstHeal](_constants_.known_proc_id.md#burstheal)
-
-## Enumeration members
-
-###  BurstHeal
-
-• **BurstHeal**: = "2"
-
-*Defined in [constants.ts:2](https://github.com/BluuArc/bfmt-utilities/blob/master/src/constants.ts#L2)*
