+++ docs/globals.md
@@ -9,6 +9,7 @@
 * ["buffs/IProcEffectFrameComposite"](modules/_buffs_iproceffectframecomposite_.md)
 * ["buffs/buff-metadata"](modules/_buffs_buff_metadata_.md)
 * ["buffs/combineEffectsAndDamageFrames"](modules/_buffs_combineeffectsanddamageframes_.md)
+* ["buffs/constants"](modules/_buffs_constants_.md)
 * ["buffs/getEffectId"](modules/_buffs_geteffectid_.md)
 * ["buffs/getEffectName"](modules/_buffs_geteffectname_.md)
 * ["buffs/getMetadataForPassive"](modules/_buffs_getmetadataforpassive_.md)
@@ -23,7 +24,6 @@
 * ["bursts/getExtraAttackDamageFramesEntry"](modules/_bursts_getextraattackdamageframesentry_.md)
 * ["bursts/getLevelEntryForBurst"](modules/_bursts_getlevelentryforburst_.md)
 * ["bursts/index"](modules/_bursts_index_.md)
-* ["constants"](modules/_constants_.md)
 * ["datamine-types"](modules/_datamine_types_.md)
 * ["extra-skills/getEffectsForExtraSkill"](modules/_extra_skills_geteffectsforextraskill_.md)
 * ["extra-skills/index"](modules/_extra_skills_index_.md)
@@ -32,7 +32,6 @@
 * ["items/index"](modules/_items_index_.md)
 * ["leader-skills/getEffectsForLeaderSkill"](modules/_leader_skills_geteffectsforleaderskill_.md)
 * ["leader-skills/index"](modules/_leader_skills_index_.md)
-* ["sp-enhancements/_constants"](modules/_sp_enhancements__constants_.md)
 * ["sp-enhancements/getAllDependenciesForSpEntry"](modules/_sp_enhancements_getalldependenciesforspentry_.md)
 * ["sp-enhancements/getAllEntriesThatDependOnSpEntry"](modules/_sp_enhancements_getallentriesthatdependonspentry_.md)
 * ["sp-enhancements/getEffectsForSpEnhancement"](modules/_sp_enhancements_geteffectsforspenhancement_.md)
