+++ docs/modules/_buffs_constants_.md
@@ -1,0 +1,10 @@
+[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["buffs/constants"](_buffs_constants_.md)
+
+# Module: "buffs/constants"
+
+## Index
+
+### Enumerations
+
+* [KNOWN_PASSIVE_ID](../enums/_buffs_constants_.known_passive_id.md)
+* [KNOWN_PROC_ID](../enums/_buffs_constants_.known_proc_id.md)
