--- docs/modules/_constants_.md
@@ -1,10 +1,0 @@
-[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["constants"](_constants_.md)
-
-# Module: "constants"
-
-## Index
-
-### Enumerations
-
-* [KNOWN_PASSIVE_ID](../enums/_constants_.known_passive_id.md)
-* [KNOWN_PROC_ID](../enums/_constants_.known_proc_id.md)
