--- docs/modules/_sp_enhancements__constants_.md
@@ -1,5 +1,0 @@
-[@bluuarc/bfmt-utilities - v0.6.0](../README.md) › [Globals](../globals.md) › ["sp-enhancements/_constants"](_sp_enhancements__constants_.md)
-
-# Module: "sp-enhancements/_constants"
-
-
