+++ typedoc.config.js
@@ -4,7 +4,7 @@ module.exports = {
 	target: 'ES5',
 	tsconfig: 'tsconfig.json',
 	readme: 'README.md',
-	exclude: 'src/index.ts',
+	exclude: ['src/index.ts', 'src/sp-enhancements/_constants.ts'],
 	theme: 'markdown',
 	out: 'docs',
 	gitRevision: 'master',
